requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
boto3==1.13.2
botocore==1.16.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.5
django-braces==1.14.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-embed-video==1.3.2
django-filter==2.2.0
django-js-asset==1.2.2
django-memcache-status==2.2
django-memcached==0.1.2
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jieba==0.39
jmespath==0.9.5
msgpack==0.6.2
msgpack-python==0.5.6
pbr==3.1.1
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.1
redis==2.10.6
requests==2.23.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
whitenoise==5.0.1
zope.interface==5.1.0
